Run stage-1 only if the output does not exist.

+++ release-tools/vm/build.sh
@@ -32,13 +32,16 @@ shift
 
 BASE_OUT_FILE="output/base/disk-vm-${VM_VERSION}.ovf"
 
-packer build -var "iso_url=${URL}" \
-             -var "iso_name=${ISO_NAME}" \
-             -var "out_file=`basename ${BASE_OUT_FILE} .ovf`" \
-             -var "vm_version=${VM_VERSION}" \
-             -machine-readable \
-             $@ 00-base.json | tee log-00.txt
+if [ ! -f "${BASE_OUT_FILE}" ]; then
 
+    packer build -var "iso_url=${URL}" \
+                 -var "iso_name=${ISO_NAME}" \
+                 -var "out_file=`basename ${BASE_OUT_FILE} .ovf`" \
+                 -var "vm_version=${VM_VERSION}" \
+                 -machine-readable \
+                 $@ 00-base.json | tee log-00.txt
+
+fi
 
 #-----------------------------------------
 # Step 2: Build AWS VM (only for releases)
